feat(client): board support listing model, dataset, runtime and job
